add eggd_athena requirements to repo
