+++ athena/bin/version.py
@@ -1,0 +1,1 @@
+VERSION = '1.5.0'
