+++ athena/data/example/example_gene_stats.tsv
@@ -1,0 +1,38 @@
+# build: GRCh37 (h37.d5)
+gene	tx	exon	min	mean	max	10x	20x	30x	50x	100x
+AIP	NM_003977.3		80	253.2745193171608	465	100.0	100.0	100.0	100.0	98.74213836478
+APC	NM_000038.5		79	291.9128985507246	649	100.0	100.0	100.0	100.0	99.558423913043
+BMPR1A	NM_004329.2		100	344.22653106102246	686	100.0	100.0	100.0	100.0	99.945024738868
+BRCA1	NM_007294.3		69	274.0322596153847	653	100.0	100.0	100.0	100.0	98.657161803714
+BRCA2	NM_000059.3		73	266.38831307413943	557	100.0	100.0	100.0	100.0	98.923633664285
+BRIP1	NM_032043.2		65	315.7521888619854	644	100.0	100.0	100.0	100.0	99.661016949153
+CDC73	NM_024529.4		90	329.4007076446281	536	100.0	100.0	100.0	100.0	99.535123966942
+CDKN1B	NM_004064.4		79	164.89778649921507	221	100.0	100.0	100.0	100.0	97.959183673469
+EPCAM	NM_002354.2		17	329.22561777777776	554	100.0	98.755555555556	91.466666666667	91.466666666667	90.133333333333
+FH	NM_000143.3		38	311.23836122331215	632	100.0	100.0	100.0	98.845931909983	94.518176572418
+GREM1	NM_013372.6		106	307.85	502	100.0	100.0	100.0	100.0	100.0
+MAX	NM_002382.4		102	277.5252658662092	503	100.0	100.0	100.0	100.0	100.0
+MEN1	NM_130799.2		12	222.09595131644312	565	100.0	98.658718330849	97.714853452558	94.734227521113	90.561351217089
+MLH1	NM_000249.3		13	344.52743115805356	737	100.0	99.886835156545	99.660505469634	99.321010939268	98.566578649566
+MSH2	NM_000251.2		19	235.5548608	453	100.0	99.872	99.616	99.392	90.656
+MSH6	NM_000179.2		38	298.7056805977118	537	100.0	100.0	100.0	99.789866915713	94.910109736166
+MUTYH	NM_001128425.1		74	245.29984263959392	406	100.0	100.0	100.0	100.0	98.527918781726
+PALB2	NM_024675.3		52	297.54906045537825	529	100.0	100.0	100.0	100.0	97.906307249411
+PMS2	NM_000535.6		0	243.039989615784	568	97.819314641745	97.403946002077	96.677050882658	93.734856351679	83.350640359986
+POLD1	NM_002691.3		39	241.63685744016652	502	100.0	100.0	100.0	99.765868886576	93.132154006243
+POLE	NM_006231.3		26	312.8382642520086	635	100.0	100.0	99.987246524678	99.668409641627	97.372784083663
+PTEN	NM_000314.6		14	290.36395833333336	509	100.0	99.353448275862	98.850574712644	98.275862068966	93.965517241379
+RAD51C	NM_058216.2		54	327.7485202135774	609	100.0	100.0	100.0	100.0	96.033562166285
+RAD51D	NM_002878.3		0	254.98551811288965	412	88.289806234204	88.289806234204	88.289806234204	88.289806234204	88.289806234204
+RET	NM_020975.4		35	264.7683417890521	632	100.0	100.0	100.0	99.412550066756	95.914552736983
+SDHA	NM_004168.3		18	297.94198692810454	522	100.0	99.912854030501	99.520697167756	98.823529411765	96.732026143791
+SDHAF2	NM_017841.2		34	186.13951807228915	321	100.0	100.0	100.0	98.450946643718	87.951807228916
+SDHB	NM_003000.2		85	327.1589032901296	712	100.0	100.0	100.0	100.0	98.803589232303
+SDHC	NM_003001.3		58	197.3551111111111	395	100.0	100.0	100.0	100.0	91.587301587302
+SDHD	NM_003002.3		21	317.81960714285714	536	100.0	100.0	98.035714285714	96.071428571429	92.321428571429
+SMAD4	NM_005359.5		52	347.71838744012774	715	100.0	100.0	100.0	100.0	96.64715274082
+STK11	NM_000455.4		62	231.00053981106615	407	100.0	100.0	100.0	100.0	97.975708502024
+TMEM127	NM_017849.3		9	173.43606177606176	306	98.584298584299	96.653796653797	94.465894465894	89.446589446589	83.783783783784
+TSC1	NM_000368.4		14	281.10650574712645	498	100.0	99.719029374202	99.540229885057	99.106002554278	97.292464878672
+TSC2	NM_000548.4		56	253.74522421524665	481	100.0	100.0	100.0	100.0	97.869955156951
+VHL	NM_000551.3		64	230.98291277258562	543	100.0	100.0	100.0	100.0	88.940809968847
