--- bin/version.py
@@ -1,1 +1,0 @@
-VERSION = '1.5.0'
